requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
requests==2.32.3
beautifulsoup4==4.12.3
anthropic==0.34.2
pydantic==2.9.2
python-multipart==0.0.9
lxml==5.3.0
